sharethought/requirements.txt
arrow==1.2.3
asgiref==3.7.2
astroid==2.15.5
async-timeout==4.0.2
backports.zoneinfo==0.2.1
binaryornot==0.4.4
black==22.10.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
coverage==7.2.7
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
dill==0.3.6
distlib==0.3.6
dj-rest-auth==4.0.1
Django==3.2
django-allauth==0.54.0
django-appconf==1.0.5
django-compressor==4.1
django-cors-headers==3.14.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-jazzmin==2.6.0
django-kn-defaults==0.3.2
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-redis==5.2.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-rq==2.6.0
django-storages==1.13.2
djangorestframework==3.14.0
djvue==0.2.1
elastic-apm==6.13.2
environ==1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==12.0.1
filelock==3.12.0
flake8==6.0.0
flake8-isort==5.0.3
gunicorn==20.0.4
identify==2.5.24
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonfield==3.1.0
lazy-object-proxy==1.9.0
limits==3.5.0
MarkupPy==1.14
MarkupSafe==2.1.2
mccabe==0.7.0
mixer==7.2.2
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.13
Pillow==9.4.0
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==2.20.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pylogbeat==2.0.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-logstash-async==2.5.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
rcssmin==1.1.0
redis==4.3.5
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.0
rq==1.15.0
sentry-sdk==1.11.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tablib==3.4.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
typing-extensions==4.6.2
urllib3==2.0.2
virtualenv==20.23.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
